Update clm_data artifact
This updates the clm_data folder to produce an artifact
that is the same as what is in /groups/esm/ClimaArtifacts/artifacts/clm_data
All the scripts used to generate the files on calhpc are copied into
the clm_data folder. Previously, create_artifacts.jl downloaded
a single surf data file. create_artifact.jl is updated to download two new files
that are used to produce the output data. create_artifact.jl also runs
several python scripts to produce the output data. The README is updated,
but information about some of the vars and files is still missing.