Update precipitation_obs artifact data and readme
The new data is from the NOAA rather than the files from
caltech data for "Error analysis of climate models...". It contains all the
same data as the old artifact, except with data for 2023-present.

The old artifact is still on box, and the new artifact has v2 appended to its name.
The new data has a different file name as well.