Add download callback to ClimaArtifactsHelper
The callback handles the condition when total size is unknown.
Total size is printed as  MB as an int if it is less than 1 GB, and GB with 2 decimal
places otherwise. The same applies to dowloaded size and rate.

The function `download_rate_callback` returns
the callback for Downloads.downloads. The callback calculates the
download rate over an interval, and prints the result at the end of
the interval. It also prints download progress. This callback
uses a closure to keep track of the time and bytes downloaded
at the end of the previous interval.